feat: Refactor server logic
Closes: #7
