+++ src/lib/mod.rs
@@ -1,4 +1,5 @@
 pub mod args;
+mod auth_server;
 pub mod authorization_code_retriever;
 pub mod authorization_code_with_pkce_retriever;
 pub mod client_credentials_retriever;
@@ -7,6 +8,5 @@ pub mod file_state;
 pub mod implicit_retriever;
 pub mod oauth_client;
 pub mod openidc_discovery;
-mod server;
 pub mod token_info;
 pub mod token_retriever;
